% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vocab.R
\name{extract_vocab}
\alias{extract_vocab}
\title{Extract OMOP Vocabularies from Source CSV Files}
\usage{
extract_vocab(vocabulary_path)
}
\arguments{
\item{vocabulary_path}{character string containing the path to the folder
containing OMOP vocabularies as CSV.}
}
\value{
a list of tables containing the vocabularies
}
\description{
Takes a single argument \code{vocabulary_path} which is the local path
to the OMOP vocabularies in CSV format. These are an essential component to
create a new OMOP database
}
