% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_xml.R
\name{extract_xml}
\alias{extract_xml}
\title{Extract data from ICNARC xml}
\usage{
extract_xml(xml_path)
}
\arguments{
\item{xml_path}{the path to a folder with ICNARC style XML. This folder
should contain NOTHING but XML. The XML should be labelled in
lexicographical order (i.e. the order you want the files to be read in).
For example, my_xml01.xml, my_xml02.xml and so on. Please do not rely on
dates in the names parse files in the correct order.}
}
\value{
a data frame with ICNARC data
}
\description{
Extract data from ICNARC xml
}
