% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cdm_tables.R
\name{setup_observation_period}
\alias{setup_observation_period}
\title{Setup Observation Period Table}
\usage{
setup_observation_period(inmem_cdm)
}
\arguments{
\item{inmem_cdm}{}
}
\description{
The observation period spans from the beggining of the study period (i.e. 1st
Jan 2014) to either today, or 60 days post death.
}
