% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cdm_tables.R
\name{setup_person_death}
\alias{setup_person_death}
\title{Setup Person/Death Table/s}
\usage{
setup_person_death(inmem_cdm, input_data, project_path)
}
\arguments{
\item{project_path}{}
}
\description{
The major difference between version 5 and 6 is that death has its own table
in 5. Thus these tables are best written out together when trying to ETL to
both.
}
