% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_xml.R
\name{transform_data_type}
\alias{transform_data_type}
\title{Transform Data Type}
\usage{
transform_data_type(input_data, input_name, lookup)
}
\arguments{
\item{input_data}{vector of data to be transformed}

\item{input_name}{name for vector to be found in lookup}

\item{lookup}{a lookup table with conversion functions}
}
\value{
a vector of the datatype specified in lookup
}
\description{
Data stored in the XML are inherently character strings. This function
transforms the raw data into the correct type, according to the ICNARC data
specification.
}
