% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify_post_code}
\alias{verify_post_code}
\title{Verify Post Code}
\usage{
verify_post_code(post_code = NULL)
}
\arguments{
\item{post_code}{a character vector of post codes}
}
\description{
Verifies that the character string provided is a valid UK postcode. This
means that it conforms to the correct specification for a UK postcode, not
that it actually exists.
}
